text_processing.py
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk import pos_tag

# Download necessary NLTK data
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('averaged_perceptron_tagger')

stop_words = set(stopwords.words('english'))

def process_text(text):
    words = word_tokenize(text)
    words = [word for word in words if word.isalnum() and word not in stop_words]
    tagged_words = pos_tag(words)
    
    # Example intent detection logic
    if "move" in words:
        return "Sure, navigating to the adjust page for you."
    elif "receive" in words:
        return "Please specify the quantity and location for the items you want to receive."
    elif "how many items" in words:
        location = extract_location(text)
        quantity = get_quantity_from_db(location)
        return f"You have {quantity} items in {location}."
    return "Sorry, I didn't understand that."

def extract_location(text):
    # Placeholder function to extract location from text
    # Implement the actual logic to extract location
    return "location X"

def get_quantity_from_db(location):
    # Placeholder function to get quantity from the database
    # Implement the actual logic to query the database
    return 500
